Make logging subsystem into a separate function and prevent restart failure upon yum reinstall
--------------------------------------

ok

aha!
i think i figured it out!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
woooooooooo hoooooooooo
why it only fails in upgrades!
when you install via yum, yum runs the postinstall script command through a subprocess, which
uses POPEN
popen creates an anonymous pipe for stdin, stdout and stderr
to communicate with yum, and relay the scriptlet output
i know this because the initrd subsystem colorizes the restart output when stdout is on a
console, and doesnt when it is on a pipe
so, in this context, instead of usnig the console, this named pipe is the stdout and stderr of
the dtc-xen process
now, the daemonize code only closes file descriptors attached to a tty
if not, doesnt close it
so when it attempts to redirect stderr, blam i fails
because its attached to a pipe without readers
hence
IOError: [Errno 9] Bad file descriptor
tadaaaaa
and thats why when you restart dtc-xen from a terminal, it works fine! because the daemonize
code corerctly detects "oh, this is a tty" and closes the respective fds